Make ParaDiag a mixin over either controller, and get it out of core
Two problems with the previous commit, both fair: the virtual ParaDiag controller derived from
its transport's controller while the MPI one did not, and `ParaDiagController` -- an algorithm --
lived in `pySDC/core/controller.py`.

Both are now the same shape, and neither is in core:

    controller_ParaDiag_nonMPI(ParaDiag, controller_nonMPI)
    controller_ParaDiag_MPI(ParaDiag, controller_MPI)

`ParaDiag` is a mixin in `implementations/controller_classes/ParaDiag.py` holding what ParaDiag
is independently of where the block sits: the parameter checks, alpha, the weighted transform,
its stages, and its answers to the three questions a driver asks. It has no `__init__` and no
controller base, so it can be mixed into either transport without the two agreeing on a
constructor. What stays with the concrete classes is only what depends on where the other steps
are -- `apply_matrix`, `prepare_Jacobians`, `compute_all_at_once_residual`, `update_G_inv`.

The MPI controller therefore drops its own `run`, `restart_block`, `spread`, `it_check` and
dispatch. That also removes a redundant collective: it was allreducing convergence itself on top
of the allreduce `CheckConvergence` already does, once per iteration.

Two seams changed with it:

- `get_active_steps` is gone; both drivers now ask `Controller.step_is_active(time, block_start,
  Tend)`, which is the same question in either transport and lets ParaDiag answer it once for
  both. Sharing `run` is what this buys.
- `update_residual_for_check` is now `prepare_convergence_check`, because the residual was only
  half of what it promised. `controller_MPI` stopped computing the end point in `it_check` when
  `send_full` took it over (#670), so a ParaDiag override that did nothing left `uend` one
  increment stale -- caught by `test_ParaDiag_MPI_matches_nonMPI`, 7.8e-10 against a 1e-14
  tolerance. The seam now names both values it makes current, and ParaDiag supplies the end
  point.

Net -260 lines. 28 configurations unchanged against master, and the two transports still agree
at rtol=1e-14.
