+++ pySDC/core/controller.py
@@ -1,7 +1,7 @@
 import logging
 import os
 import sys
-from typing import Any, Dict, List, Optional, Type, Union
+from typing import Any, Dict, List, Optional, Type
 import numpy as np
 
 from pySDC.core.base_transfer import BaseTransfer
@@ -336,6 +336,24 @@ class Controller(object):
                         f'step(s). Use a preconditioner with fixed coefficients, e.g. MIN-SR-S or LU.'
                     )
 
+    def step_is_active(self, time: float, block_start: float, Tend: float) -> bool:
+        """
+        Whether a step starting at `time`, in a block starting at `block_start`, still has work.
+
+        The default is that a step runs if it starts before the end of the interval, so a block
+        may be run partially. An algorithm that couples its steps too tightly to drop one of them
+        answers this from `block_start` instead and runs the block whole.
+
+        Args:
+            time (float): when this step starts
+            block_start (float): when the first step of this step's block starts
+            Tend (float): ending time
+
+        Returns:
+            bool: whether this step takes part
+        """
+        return time < Tend - 10 * np.finfo(float).eps
+
     def setup_convergence_controllers(self, description: Dict[str, Any]) -> None:
         '''
         Setup variables needed for convergence controllers, notably a list containing all of them and a list containing
@@ -431,137 +449,3 @@ class Controller(object):
         for hook in self.hooks:
             stats = {**stats, **hook.return_stats()}
         return stats
-
-
-class ParaDiagController(Controller):
-    """
-    What a ParaDiag controller needs on top of the controller it inherits its driver from.
-
-    This carries no `__init__` of its own. A concrete ParaDiag controller calls
-    `prepare_ParaDiag_params` on the two dictionaries and then hands them to the driver's
-    initialisation, so that ParaDiag can be mixed into either transport's controller without the
-    two having to agree on a constructor signature.
-    """
-
-    @staticmethod
-    def prepare_ParaDiag_params(controller_params: Dict[str, Any], description: Dict[str, Any]) -> None:
-        """
-        Check and complete the parameters ParaDiag needs, in place.
-
-        Call this *before* the controller's own initialisation: it only reads and writes the two
-        dictionaries, and must have run by the time the steps are built.
-
-        Args:
-            controller_params (dict): parameter set for the controller and the steps
-            description (dict): all the parameters to set up the rest (levels, problems, ...)
-        """
-        from pySDC.implementations.sweeper_classes.ParaDiagSweepers import QDiagonalization
-
-        if QDiagonalization in description['sweeper_class'].__mro__:
-            description['sweeper_params']['ignore_ic'] = True
-            description['sweeper_params']['update_f_evals'] = False
-        else:
-            logging.getLogger('controller').warning(
-                f'Warning: Your sweeper class {description["sweeper_class"]} is not derived from {QDiagonalization}. You probably want to use another sweeper class.'
-            )
-
-        if not controller_params.get('all_to_done', True):
-            raise NotImplementedError('ParaDiag only implemented with option `all_to_done=True`')
-        if 'alpha' not in controller_params.keys():
-            from pySDC.core.errors import ParameterError
-
-            raise ParameterError('Please supply alpha as a parameter to the ParaDiag controller!')
-        controller_params['average_jacobian'] = controller_params.get('average_jacobian', True)
-
-        controller_params['all_to_done'] = True
-
-    @staticmethod
-    def resolve_alpha(alpha: Any, k: int = 0) -> float:
-        """
-        Read the alpha for iteration `k` out of whatever the user supplied.
-
-        `alpha` may be a single number, a sequence indexed by iteration (the last entry is reused once
-        it runs out), or a callable taking the iteration index. Making it iteration dependent lets the
-        outer iteration start with a well-conditioned alpha and tighten it later.
-
-        Static because the steps need an alpha before the controller has parameters to read it from.
-
-        Args:
-            alpha: the alpha parameter as supplied by the user
-            k (int): iteration index
-
-        Returns:
-            float: alpha to use for this iteration
-        """
-        if callable(alpha):
-            return float(alpha(k))
-        if hasattr(alpha, '__len__'):
-            return float(alpha[min(k, len(alpha) - 1)])
-        return float(alpha)
-
-    def get_alpha(self, k: int = 0) -> float:
-        """
-        Get the ParaDiag alpha parameter for iteration `k`.
-
-        Args:
-            k (int): iteration index
-
-        Returns:
-            float: alpha to use for this iteration
-        """
-        return self.resolve_alpha(self.params.alpha, k)
-
-    def get_FFT_matrices(self, k: int = 0) -> Any:
-        """
-        Get the weighted FFT and iFFT matrices for iteration `k`, rebuilding them only when alpha
-        actually changes.
-
-        Args:
-            k (int): iteration index
-
-        Returns:
-            tuple: the forward and backward weighted FFT matrices
-        """
-        alpha = self.get_alpha(k)
-        if getattr(self, '_cached_alpha', None) != alpha:
-            from pySDC.helpers.ParaDiagHelper import get_weighted_FFT_matrix, get_weighted_iFFT_matrix
-
-            self._FFT_matrix = get_weighted_FFT_matrix(self.n_steps, alpha)
-            self._iFFT_matrix = get_weighted_iFFT_matrix(self.n_steps, alpha)
-            self._cached_alpha = alpha
-        return self._FFT_matrix, self._iFFT_matrix
-
-    def update_G_inv(self, k: int = 0) -> None:
-        """
-        Rebuild G^-1 on the local step(s) when alpha changes with the iteration.
-
-        G^-1 depends on alpha, so an iteration dependent alpha means the sweeper's diagonalization has
-        to be recomputed. Subclasses implement this because only they know which steps they own.
-
-        Args:
-            k (int): iteration index
-        """
-        raise NotImplementedError('ParaDiag controllers have to implement update_G_inv')
-
-    def FFT_in_time(self, quantity: Any, k: int = 0) -> None:
-        """
-        Compute weighted forward FFT in time. The weighting is determined by the alpha parameter in ParaDiag
-
-        Note: The implementation via matrix-vector multiplication may be inefficient and less stable compared to an FFT
-              with transposes!
-
-        Args:
-            quantity (str): the level attribute to transform
-            k (int): iteration index, for an iteration dependent alpha
-        """
-        self.apply_matrix(self.get_FFT_matrices(k)[0], quantity)
-
-    def iFFT_in_time(self, quantity: Any, k: int = 0) -> None:
-        """
-        Compute weighted backward FFT in time. The weighting is determined by the alpha parameter in ParaDiag
-
-        Args:
-            quantity (str): the level attribute to transform
-            k (int): iteration index, for an iteration dependent alpha
-        """
-        self.apply_matrix(self.get_FFT_matrices(k)[1], quantity)
